Refactor MainLayout component
Refactored the main layout component, moving the footer to a separate file.
